publish a sp object to geoserver with multiple wms stlying information

publish a sp object to geoserver with multiple wms stlying information

utils.publishSP2GeoServerWithMultiStyles(
  spobj,
  layerprefix = "my_",
  styleprefix = "my_stl_",
  attrname_vec = c(""),
  layerdisplyname_vec = c(""),
  palettename_vec = c("Reds"),
  colorreverseorder_vec = c(FALSE),
  colornum_vec = c(5),
  classifier_vec = c("Jenks"),
  bordercolor_vec = c("black"),
  borderwidth_vec = c(1),
  bordervisible_vec = c(TRUE),
  styletype_vec = c("single")
)

Arguments

spobj
a sp object. IMPORTANT: Each column of spobj MUST NOT have attribute name exceeding length (int)10, failed to do so may result in NO Style being generated on GeoServer

layerprefix
a prefix for layer name, will make the generated data layers in GeoServer more identifiable

styleprefix
a prefix for style name, will make the generated styles in GeoServer more identifiable

attrname_vec
a vector of attribute names that the style will be created for

palettename_vec
a vector of palette names. check colors defined under each palette name at http://colorbrewer2.org acceptable color PaletteNames are: YlOrRd: 9 colors PRGn: 11 colors PuOr:11 colors RdGy: 11 colors Spectral: 11 colors Grays: 9 colors PuBuGn: 9 colors RdPu: 9 colors BuPu: 9 colors YlOrBr: 9 colors Greens: 9 colors BuGn: 9 colors Accents: 8 colors GnBu: 9 colors PuRd: 9 colors Purples: 9 colors RdYlGn: 11 colors Paired: 12 colors Blues: 9 colors RdBu: 11 colors Oranges: 9 colors RdYlBu: 11 colors PuBu: 9 colors OrRd: 9 colors Set3: 12 colors Set2: 8 colors Set1: 9 colors Reds: 9 colors PiYG: 11 colors Dark2: 8 colors YlGn: 9 colors BrBG: 11 colors YlGnBu: 9 colors Pastel2: 8 colors Pastel1: 9 colors

colorreverseorder_vec
a vector of color reverse order. true or false(default), whether to reverse the color order defined in a palette

colornum_vec
a vector of the number of colors to be classified. default 5. in range of 2-15

classifier_vec
a vector of classification function name. acceptable classifier function names are: Jenks(default), EqualInterval, Quantile, StandardDeviation

bordercolor_vec
a vector of border color for polygon layers, default "black", can be white, gray or black.

borderwidth_vec
a vector of border stroke width (integer value) for polygon layers. in range 1, 10, default 1

bordervisible_vec
a vector of border visibility for polygon layers. TRUE/FALSE, default, TRUE

styletype_vec
a vector of styling type, default "single"

displayname_vec
a vector of readable display names for data layer

Value

a list elements to be included in geolayers